chore: move EventCard into pages/Events
