+++ src/pages/Events/Content/EventsList/EventsList.tsx
@@ -3,7 +3,7 @@ import type { IEvent, IEventDB } from 'src/models/events.models'
 import { Button } from 'oa-components'
 import { Flex, Box, Link, Heading } from 'theme-ui'
 import MoreContainer from 'src/components/MoreContainer/MoreContainer'
-import EventCard from 'src/components/EventCard/EventCard'
+import EventCard from 'src/pages/Events/EventCard/EventCard'
 import TagsSelect from 'src/components/Tags/TagsSelect'
 import { inject, observer } from 'mobx-react'
 import type { EventStore } from 'src/stores/Events/events.store'
